Created t9 engine object on separate file. Docstring fixes
